Separating Upgrade test and ClickHouse Migration Tests
Latest changes in Antrea CRDs require us to delete the downgrade and check test, to keep it on par with Antrea, we have separated the upgrade test and the ClickHouse migration tests.
In order to test for the data Schema management, we have added new ClickHouse migration test.
