+++ .github/workflows/kind.yml
@@ -222,6 +222,63 @@ jobs:
           name: upgrade-from-theia-version-n-1.tar.gz
           path: log.tar.gz
           retention-days: 30
+  
+  test-ClickHouse-migrate-from-N-1:
+    name: Migrate ClickHouse from Theia version N-1
+    needs: 
+      - build-clickhouse-monitor-image
+      - build-clickhouse-server-image 
+    runs-on: [ubuntu-latest]
+    steps:
+      - name: Free disk space
+        # https://github.com/actions/virtual-environments/issues/709
+        run: |
+          sudo apt-get clean
+          sudo rm -rf /usr/share/dotnet
+          sudo rm -rf /opt/ghc
+          sudo rm -rf "/usr/local/share/boost"
+          sudo rm -rf "$AGENT_TOOLSDIRECTORY"
+          sudo rm -rf "/usr/local/lib/android"
+          df -h
+      - uses: actions/checkout@v3
+      - uses: actions/setup-go@v4
+        with:
+          go-version-file: 'go.mod'
+      - name: Download ClickHouse monitor images from previous jobs
+        uses: actions/download-artifact@v3
+        with:
+          name: clickhouse-monitor
+      - name: Load Theia image
+        run:  |
+          docker load -i clickhouse-monitor.tar
+          docker tag antrea/theia-clickhouse-monitor:latest projects.registry.vmware.com/antrea/theia-clickhouse-monitor:latest
+      - name: Download ClickHouse server images from previous jobs
+        uses: actions/download-artifact@v3
+        with:
+          name: clickhouse-server
+      - name: Load Theia image
+        run:  |
+          docker load -i clickhouse-server.tar
+          docker tag antrea/theia-clickhouse-server:latest projects.registry.vmware.com/antrea/theia-clickhouse-server:latest
+      - name: Install Kind
+        run: |
+          curl -Lo ./kind https://github.com/kubernetes-sigs/kind/releases/download/${KIND_VERSION}/kind-$(uname)-amd64
+          chmod +x ./kind
+          sudo mv kind /usr/local/bin
+      - name: Run migrate test
+        run: |
+          mkdir log
+          ANTREA_LOG_DIR=$PWD/log ./ci/kind/test-clickhouse-migrate.sh  --from-version-n-minus 1
+      - name: Tar log files
+        if: ${{ failure() }}
+        run: tar -czf log.tar.gz log
+      - name: Upload test log
+        uses: actions/upload-artifact@v3
+        if: ${{ failure() }}
+        with:
+          name: migrate-clickhouse-from-theia-version-n-1.tar.gz
+          path: log.tar.gz
+          retention-days: 30
 
   # Runs after all other jobs in the workflow succeed and deletes Theia Docker images uploaded as temporary
   # artifacts. It uses a third-party, MIT-licensed action (geekyeggo/delete-artifact). While Github
@@ -236,6 +293,7 @@ jobs:
     - build-theia-manager-image
     - test-e2e-encap
     - test-upgrade-from-N-1
+    - test-ClickHouse-migrate-from-N-1
     runs-on: [ubuntu-latest]
     steps:
     - name: Delete spark-jobs
