+++ test/e2e/theia_clickhouse_test.go
@@ -53,7 +53,7 @@ var tableColumnNumberMap = map[string]string{
 }
 
 func TestTheiaClickHouseStatusCommand(t *testing.T) {
-	config := FlowVisibiltiySetUpConfig{
+	config := FlowVisibilitySetUpConfig{
 		withSparkOperator:     false,
 		withGrafana:           false,
 		withClickHouseLocalPv: false,
