+++ test/e2e/throughputanomalydetection_test.go
@@ -39,7 +39,7 @@ const (
 )
 
 func TestAnomalyDetection(t *testing.T) {
-	config := FlowVisibiltiySetUpConfig{
+	config := FlowVisibilitySetUpConfig{
 		withSparkOperator:     true,
 		withGrafana:           false,
 		withClickHouseLocalPv: false,
